Build a kerning table on font creation
This allows freetype references to be removed from glyph and font
data structures and paves the way for writing rasterized font's to
disk. This could allow an application to avoid repeatedly computing
glyph signed distance fields. It also makes it possible to have
"deployment" builds that don't need to link to freetype to render text.