First pass at handling unicode characters
Interpret charset as a UTF-8 string and convert to UTF-32 during
font creation. This has only been tested on Linux, there will probably
need to be fixes for Windows. The approach to glyph lookup will probably
have to be changed since it is now possible that the glyph array
will need to store up to 2^32 entries.