Avoid using 'glplatform' in file names
I'm anticipating renaming the project, perhaps more than once, so having
the project name in files is an obsticle. This change required a
glbindify update to 0.5.1 to remove the "glplatform-" prefix from the files.
Renaming the files was not sufficient since the sources includes the headers
which would have required modifying the generated sources without a glbindify
update.